refactor: factorize fill with zeroes method
