+++ api/services/trip/src/actions/PublicStatsAction.ts
@@ -5,6 +5,8 @@ import { ResultInterface } from '../shared/trip/stats.contract';
 import { TripRepositoryProvider } from '../providers/TripRepositoryProvider';
 import { StatCacheRepositoryProviderInterfaceResolver } from '../interfaces/StatCacheRepositoryProviderInterface';
 import { PublicTripSearchInterface } from '../shared/trip/common/interfaces/TripSearchInterface';
+import { fillWithZeroes } from '../helpers/fillWithZeroesHelper';
+import { StatInterface } from '../interfaces/StatInterface';
 
 @handler({
   service: 'trip',
@@ -16,6 +18,11 @@ export class PublicStatsAction extends Action {
   }
 
   public async handle(params: PublicTripSearchInterface): Promise<ResultInterface> {
-    return (await this.cache.getOrBuild(async () => this.pg.stats(params), params)) || [];
+    return (
+      (await this.cache.getOrBuild(async () => {
+        const statInterface: StatInterface[] = await this.pg.stats(params);
+        return statInterface.length === 0 ? [] : fillWithZeroes(statInterface, params);
+      }, params)) || []
+    );
   }
 }
